RIP-547, admin users get run-job options and job history on landing page
